MapRenderer: Simplified parameters to RenderTileCallback
By moving more code out of the function determining the tile and screen
positions, and into the callback and its context, we reduce the amount
of parameters needing to go into the callback and increase the
flexibility.

Based on this API we could potentially implement a "render by tile
column" option, though that would still require quite some work in other
areas.